float remplaced by double, PCMBufferFormat replace AudioLineFormat, Markers in WAV, FFT, RMSCalculator, First version of SynthRipper

+++ audio-workshop/src/main/java/com/hypercube/workshop/audioworkshop/common/pcm/PCMtoSampleFunction.java
@@ -4,5 +4,5 @@ import java.nio.ByteBuffer;
 
 @FunctionalInterface
 public interface PCMtoSampleFunction {
-    void convert(ByteBuffer pcmBuffer, float[][] normalizedInput, int nbSamples, int nbChannels);
+    void convert(ByteBuffer pcmBuffer, double[][] normalizedInput, int nbSamples, int nbChannels);
 }
